requirement.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
